impl fraudclient + removing duplicates

+++ customer/src/main/java/com/sergiocuacor/microservices/customer/Customer.java
@@ -27,7 +27,7 @@ public class Customer {
     }
 
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
